+++ scripts/shared/globals.js
@@ -3,4 +3,5 @@ export const MODULE_NAME = "banter-4e-modifications";
 export const ENABLE_ACTIVE_DEFENSE = "enable-active-defense";
 export const ENABLE_MARK_OWNERSHIP = "enable-mark-ownership";
 export const ENABLE_TRIGGER_PROMPTS = "enable-trigger-prompts";
+export const ENABLE_OPPORTUNITY_ATTACK_CHAT_ACTIONS = "enable-opportunity-attack-chat-actions";
 export const ENABLE_DEBUG_LOGGING = "enable-debug-logging";